requirements.txt

Pillow
torch
torchvision
opencv-python
